admin/requirements.txt
# generated by rye
# use `rye lock` or `rye sync` to update this lockfile
#
# last locked with the following flags:
#   pre: false
#   features: []
#   all-features: false
#   with-sources: false

aiosqlite==0.17.0
    # via tortoise-orm
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via httpx
    # via starlette
apscheduler==3.10.4
    # via portr-admin
asyncpg==0.29.0
    # via tortoise-orm
certifi==2024.2.2
    # via httpcore
    # via httpx
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.1.0.post1
    # via portr-admin
fastapi==0.109.2
    # via portr-admin
h11==0.14.0
    # via httpcore
    # via uvicorn
httpcore==1.0.3
    # via httpx
httpx==0.26.0
    # via portr-admin
idna==3.6
    # via anyio
    # via email-validator
    # via httpx
iso8601==1.1.0
    # via tortoise-orm
jinja2==3.1.3
    # via portr-admin
markupsafe==2.1.5
    # via jinja2
nanoid==2.0.0
    # via portr-admin
pydantic==2.6.1
    # via fastapi
    # via portr-admin
    # via pydantic-settings
pydantic-core==2.16.2
    # via pydantic
pydantic-settings==2.2.1
    # via portr-admin
pypika-tortoise==0.1.6
    # via tortoise-orm
python-dotenv==1.0.1
    # via pydantic-settings
python-slugify==8.0.4
    # via portr-admin
python-ulid==2.2.0
    # via portr-admin
pytz==2024.1
    # via apscheduler
    # via tortoise-orm
six==1.16.0
    # via apscheduler
sniffio==1.3.0
    # via anyio
    # via httpx
starlette==0.36.3
    # via fastapi
text-unidecode==1.3
    # via python-slugify
tortoise-orm==0.20.0
    # via portr-admin
    # via aiosqlite
    # via fastapi
    # via pydantic
    # via pydantic-core
tzlocal==5.2
    # via apscheduler
unidecode==1.3.8
    # via python-slugify
uvicorn==0.27.1
    # via portr-admin
